refactor: isolate skill authoring reports
